450: Preventing the codegen from using JsonNullable
Currently, the codegen produces output which includes an unused import for JsonNullable. This causes issues as this class is not on the classpath and therefore causes compilation errors.

Disabling support for JsonNullable as we don't use this functionality and it seems unlikely that we would want to in the future.

Issue: SecureBankingAccessToolkit/SecureBankingAccessToolkit#450