795: Adding testdata spring config which supports configuring the AccountIds generated by the FakeDataApiController (#129)
* 795: Adding testdata spring config which supports configuring the AccountIds generated by the FakeDataApiController

New configuration allows test AccountIds to be configured for usernames. When FakeDataApiController generates test data for a user, it will check if there are preconfigured AccountIds and use these in order they appear in the config file.

The configuration is optional. It may be completely omitted, or have no configuration for a particular username. In this case the AccountIds are generated randomly. For usernames with configuration, random AccountIds will also be produced when the configured list of ids is exhausted.

Making this change allows us to have random test data with predictable AccountIds, which makes testing easier by allowing us to specify known AccountIds in requests which require a DebtorAccount to be supplied.

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/795