592: idempotency support (#128)
* 592: idempotency support
- Added idempotency key request header in vrp payment submission
- Added idempotency key validation in vrp payment submission
Issue: https://github.com/secureapigateway/secureapigateway/issues/592

* Fix vrp controller constructor

* Fix conflicts rebase master

* 795: Adding testdata spring config which supports configuring the AccountIds generated by the FakeDataApiController (#129)

* 795: Adding testdata spring config which supports configuring the AccountIds generated by the FakeDataApiController

New configuration allows test AccountIds to be configured for usernames. When FakeDataApiController generates test data for a user, it will check if there are preconfigured AccountIds and use these in order they appear in the config file.

The configuration is optional. It may be completely omitted, or have no configuration for a particular username. In this case the AccountIds are generated randomly. For usernames with configuration, random AccountIds will also be produced when the configured list of ids is exhausted.

Making this change allows us to have random test data with predictable AccountIds, which makes testing easier by allowing us to specify known AccountIds in requests which require a DebtorAccount to be supplied.

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/795

* 797: Removing RS ingress (#130)

All calls to RS must go via IG (and the IG ingress).

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/797

* 239: Call IG internal svc address to access consent repo (#131)

The IG Consent Repo (/repo route), is an internal feature. Use the internal IG address to access it, so that the ingress rule can be removed.

Note: In future the /repo route will be removed in favour of a dedicated consent API

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/239

* Fix conflicts rebase master

---------

Co-authored-by: Dan Badham <106093201+dbadham-fr@users.noreply.github.com>