874: Fixing self links for PaymentFundsConfirmationApiController
The existing implementation for the self links is broken as it includes the /backoffice API in the path, it is also

All funds confirmation requests get mapped on to the single PaymentFundsConfirmationApiController (that is part of the /backoffice API). In order to build the self links, this controller needs to know that original OB URL that was used to call IG.

Adding `x-ob-url` header parameter and using it to build the self links. This is the same pattern used by the Accounts API.

https://github.com/SecureApiGateway/SecureApiGateway/issues/874