888: Allowing spring config to be overridden by helm values
Deployment specific Spring configuration can be provided by adding the properties as yaml to the springConfig object in the helm values file.

If Spring Config is provided in this manner, then a ConfigMap is created which contains the configuration block. This is then mounted into the RS docker container under /app/config/application.yml. Spring property resolution will check for config in this file before it falls back to the application.yml found within the JAR.

Removing the testdata deployment specific config from applicaiton.yml as this is FR deployment specific; for dev this will be supplied via ArgoCD. Adding the configuration to application-test.yml as unit tests use it.

https://github.com/SecureApiGateway/SecureApiGateway/issues/888