580: Initial commit of Validation Framework
Adding new module: secure-api-gateway-ob-uk-rs-validation which has core and obie submodules.

The core module contains the validation interfaces and classes for the framework. This module has no dependencies on other rs modules or open banking modules.

The obie module contains Open Banking specific implementations of the framework. This module depends on the core module and secure-api-gateway-ob-uk-common-* modules. Some example validator implementations are provided.

Committing this now to ensure it builds in the CI pipeline, and to share the outline of the solution with other developers.

https://github.com/SecureApiGateway/SecureApiGateway/issues/580