981: Invalid file payment BigDecimal equiality verification
The IG and IDM components that deserialize/serialize and stores the consent as json object uses a mapper where the BigDecimals value differs from the request when a request is send to the RS backend.
The initiation equality verification fails because the bigDecimals between the initiation objects differs from submit payment request and consent retrieved from IDM where is stored in different precision format.

To cope that issue is provided a custom deserialization and serialization for the backend services to deserialize the BigDecimals object from the json payload always in the same way.
So that the equality verification of the initiation sections between payment submission and consent it's performed using the same BigDecimal precision format, either the request payload or the payload retrieved from IDM.
This issue should go away when the managing of consents has been moved to RS

- Added BigDecimal deserializer and Serializer
- Added Unit tests coverage
- Bumped commons release 1.0.3
- Updated Spring custom object mapper
- Deleted custom rs object mapper to use the bean provided by Spring context
- Upgrade payment consent calculate elements as a Service to be injected in the controllers
- Updated calculate elements controllers for payments

Issue: https://github.com/secureapigateway/secureapigateway/issues/981