991: Creating Spring Boot module for java module secure-api-gateway-ob-uk-rs-cloud-client
CloudClientModuleConfiguration controls the package scan for the cloud-client module, restricting it only to the subpackages in the java module.

Updated RSServerApplication to remove the broad component scan and rely on the default component scan done by @SpringBootApplication annotation, which uses the annotated class's package as the base package.

Added an @Import for the cloud client module to RSServerApplication

https://github.com/SecureApiGateway/SecureApiGateway/issues/991