984: Adding DomesticVrpConsentsApi and Controller
We need the DomesticVrpConsentsApi (and Controller) in order to support the `funds-confirmation` API Operation. Other payments APIs reuse a backoffice controller for this purpose, however VRP request is more complex and it makes sense to have its own controller.

As we do not currently support Consent APIs in this component, we have removed the other API operations from the generated code; only the funds confirmations functionality is available.

The DomesticVrpConsentsApiController is just a generated stub at the moment (which throws UnsupportedOperationException), real implementation to be provided.

DomesticVrpConsentsApiControllerTest integration test which can be used to call the API.

https://github.com/SecureApiGateway/SecureApiGateway/issues/984