984: Adding BasePaymentRequestValidator and example implementation for Domestic Payments
BasePaymentRequestValidator provides common validation functionality for payment requests. It handles validating that risk and initiation match between the payment consent and the payment request.

OBWriteDomestic2Validator provides a concrete impl of BasePaymentRequestValidator. Note: no specific validation rules have been implemented for this payment type, it is provided as an example to be enhanced at a later date.

Removed PaymentConsentValidator, the requirements for consent validator are not concrete at the moment. Removing this base class as it did not provide any useful functionality.

https://github.com/SecureApiGateway/SecureApiGateway/issues/984