984: Adding DefaultOBValidationModule
Default validation configuration creates beans that can be wired into controllers to do validation of requests. Added validation to the DomesticPaymentsApiController which demonstrates how the framework can be used.

DomesticPaymentsApiController is a simple use case, only validation that the risk and initiation match with the consent is implemented. The OBWriteDomestic2Validator can be extended to apply more validation rules, and the validation service dependency in the controller allows validation to be customised by the customer.

https://github.com/SecureApiGateway/SecureApiGateway/issues/984