580: Validation Framework (#178)
* 580: Initial commit of Validation Framework

Adding new module: secure-api-gateway-ob-uk-rs-validation which has core and obie submodules.

The core module contains the validation interfaces and classes for the framework. This module has no dependencies on other rs modules or open banking modules.

The obie module contains Open Banking specific implementations of the framework. This module depends on the core module and secure-api-gateway-ob-uk-common-* modules. Some example validator implementations are provided.

Committing this now to ensure it builds in the CI pipeline, and to share the outline of the solution with other developers.

https://github.com/SecureApiGateway/SecureApiGateway/issues/580

* 580: Adding OBValidationService

OBValidationService provides a means of applying validation and raising exceptions if validation fails. This can be reused across OB services, providing a standard way for raising exceptions.

https://github.com/SecureApiGateway/SecureApiGateway/issues/580

* 984: Adding BasePaymentRequestValidator and example implementation for Domestic Payments

BasePaymentRequestValidator provides common validation functionality for payment requests. It handles validating that risk and initiation match between the payment consent and the payment request.

OBWriteDomestic2Validator provides a concrete impl of BasePaymentRequestValidator. Note: no specific validation rules have been implemented for this payment type, it is provided as an example to be enhanced at a later date.

* 984: Adding DefaultOBValidationModule

Default validation configuration creates beans that can be wired into controllers to do validation of requests. Added validation to the DomesticPaymentsApiController which demonstrates how the framework can be used.

DomesticPaymentsApiController is a simple use case, only validation that the risk and initiation match with the consent is implemented. The OBWriteDomestic2Validator can be extended to apply more validation rules, and the validation service dependency in the controller allows validation to be customised by the customer.

https://github.com/SecureApiGateway/SecureApiGateway/issues/580