958: Implementing simple idempotency mechanism for payment submission
- Fixing bug in PaymentSubmissionRepository which meant that certain payments do not find any records because they query the wrong field
  - Created Extension repository definitions and implementations for each payment type which use MongoTemplate directly to do the querying
- Creating IdempotentPaymentService which is used to find any existing payments with matching idempotency data
- Idempotency checks are shifted earlier in the controller impl in order to avoid validation failures

As noted in the code, this approach to idempotency is simplified. It is not guaranteed to work when there are concurrent requests, due to doing a check to see if a payment exists without any locking. This approach is deemed good enough for a test facility, as it allows developers to test the contract in a sequential fashion.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958