958: Moving the idempotent payment save logic into IdempotentPaymentService
SinglePaymentForConsentIdempotentPaymentService impl uses insert (rather than save) to create new payment record. As these payments use the consentId as the primary key then we can prevent duplicate payments, DuplicateKeyException can be handled when the insert fails and the existing record can be returned if the idempotency data matches

VRPIdempotentPaymentService impl uses save as before. As there can be multiple payments then a unique id is generated for the primary key, therefore we are unable to use a key constraint to prevent duplicates.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958