1089: Fixing concurrency issues and memory leak in the Discovery API (#223)
Initialising the AvailableApiEndpoint collection and blacklisted API collection at startup, as this data does not change then it can be retrieved once and stored. This fixes the memory leak and ConcurrentModificationExceptions by only building these collections once.

Changed AvailableApiEndpoint object to store a uriPath rather than a full uri that can be used externally, which removes any dynamic data from this object enabling us to do the processing once. A fully qualified externally callable uri is then built in the DiscoveryController using the request context data i.e. X-Forwarded-Host + X-Forwarded-Path.

In order to initialise the blacklist at startup, bean dependencies needed to be adjusted which resulted in a cyclic dependency on the Spring WebMvc. This issue has been resolved by doing late binding of the DiscoveryApiService in the DisabledEndpointInterceptor by using the new ApplicationStartupListener component.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1089