958: Removing PaymentSubmissionRepository and extension repos
As we use the convention of setting the payment.id = consentId then there is no need for custom queries, we can query the regular id field as supported by the MongoRepository.

The added benefit of this approach is that the payment.id field is indexed so queries will be much faster than searching inside the nested OB datamodel object.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958