958: Implementing simple idempotency mechanism for payment submission (#221)
* 958: Implementing simple idempotency mechanism for payment submission

- Fixing bug in PaymentSubmissionRepository which meant that certain payments do not find any records because they query the wrong field
- Creating IdempotentPaymentService which is used to find any existing payments with matching idempotency data
- Idempotency checks are shifted earlier in the controller impl in order to avoid validation failures

As noted in the code, this approach to idempotency is simplified. It is not guaranteed to work when there are concurrent requests, due to doing a check to see if a payment exists without any locking. This approach is deemed good enough for a test facility, as it allows developers to test the contract in a sequential fashion.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958

* 958: Marking the IdempotentRepositoryAdapter as deprecated

Continuing to use IdempotentRepositoryAdapter in legacy code.

Only controllers that are used for v3.1.10 API will migrate to the new service.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958

* 958: Moving the idempotent payment save logic into IdempotentPaymentService

SinglePaymentForConsentIdempotentPaymentService impl uses insert (rather than save) to create new payment record. As these payments use the consentId as the primary key then we can prevent duplicate payments, DuplicateKeyException can be handled when the insert fails and the existing record can be returned if the idempotency data matches

VRPIdempotentPaymentService impl uses save as before. As there can be multiple payments then a unique id is generated for the primary key, therefore we are unable to use a key constraint to prevent duplicates.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958

* 958: Removing PaymentSubmissionRepository

As we use the convention of setting the payment.id = consentId then there is no need for custom queries, we can query the regular id field as supported by the MongoRepository.

The added benefit of this approach is that the payment.id field is indexed so queries will be much faster than searching inside the nested OB datamodel object.

https://github.com/SecureApiGateway/SecureApiGateway/issues/958