1107: Improving GlobalExceptionHandler logging (#236)
* 1107: Improving GlobalExceptionHandler logging for OBErrorResponseException

OBErrorResponseException does not set a message on the super class, therefore logging the message always returns null. Now we log the important fields within the OBErrorResponseException, which includes a list of error objects, to build a more informative message

https://github.com/SecureApiGateway/SecureApiGateway/issues/1107

* 1107: Cleanup GlobalExceptionHandler logging

Removed unused InvalidConsentException + Handler

Cleaned up logging to log the x-fapi-interaction-id.

Created a RuntimeException handler method: handleUnexpectedRuntimeException , which handles RuntimeExceptions for which a more specific handler does not exist. These exceptions are unexpected, so are logged as errors with the stacktrace

https://github.com/SecureApiGateway/SecureApiGateway/issues/1107