1258: Specifying media types that are produced and/or consumed by the DataApi
When the media types supported are not specified, then Spring will do content negotiation with the client and attempt to work with the most appropriate type.

Our APIs are JSON based and should be configured as such. A benefit of this change is that static analysis tools, like CodeQL, will not flag potential XSS attack as being possible on this API.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1258