1430: Fix party not found when updating user data
The DataUpdater has been fixed to query for the FRParty object using the userId.

DataApiController has been updated to ensure that the userId field is always set. The update method is currently keyed off the username field, any value in the userId field will be overwritten to the value retrieved from the platform for the username.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1430