1430: Fix Data Export bug
A bug in the data export was causing the exported data for an individual user to contain data for all of the users.

The bug was due to the data paging being implemented incorrectly.

New service: DataExporter has been created which handles gathering the account data. It contains a reusable function which handles processing a paging query.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1430