1430: Remove corrupt Standing Order and Scheduled Payments from data export
DataExporter now supports a filter which can be used to remove data from the export.

For Standing Order and Scheduled Payments, this filter removes corrupt data. That is, data which cannot be updated by the DataUpdater due to the document id not matching the OB id (nested standingOrderId or scheduledPaymentId field).

As the export returns OB objects, then the OB id is used in the queries to find the database object to update. Therefore, the values need to match.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1430