1430: Stop creating Accounts API data when submitting payments (#323)
* 1430: Remove corrupt Standing Order and Scheduled Payments from data export

DataExporter now supports a filter which can be used to remove data from the export.

For Standing Order and Scheduled Payments, this filter removes corrupt data. That is, data which cannot be updated by the DataUpdater due to the document id not matching the OB id (nested standingOrderId or scheduledPaymentId field).

As the export returns OB objects, then the OB id is used in the queries to find the database object to update. Therefore, the values need to match.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1430

* 1430: Stop creating Account API data when payments are submitted

Submitting Scheduled Payments & Standing Orders via the OB Payments API no longer creates data that can be retrieved via the Accounts API.

The data being created for the Accounts API in this way was corrupt, and caused issues when attempting to update the existing static data via the Data API.

Upon reviewing this functionality, the decision has been made that the Accounts API should return static data. This data is either supplied by the Data API or it is randomly generated using the Fake Data API.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1430