1479: Remove accountIdentifierName from account queries
The accountIdentifierName corresponds to the name of the user that owns the account.

When querying by account identifiers then the userId, identification value and schemeName must be provided. The name is optional in the schema and must match the name associated with the userId, so is superfluous in this case.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1479